パーミッションに関する処理をPermissionクラスに任せる
