+++ 08.ls_object/ls_command_oop.rb
@@ -6,5 +6,6 @@ require 'etc'
 require 'date'
 require_relative 'file_data'
 require_relative 'calculation'
+require_relative 'permission'
 
 FileData.output
